move naming classes to Doctrine\ORM\Mapping
