+++ lib/Doctrine/ORM/Mapping/ClassMetadataInfo.php
@@ -21,8 +21,6 @@ namespace Doctrine\ORM\Mapping;
 
 use Doctrine\DBAL\Types\Type;
 use ReflectionClass;
-use Doctrine\ORM\DefaultNamingStrategy;
-use Doctrine\ORM\NamingStrategy;
 
 /**
  * A <tt>ClassMetadata</tt> instance holds all the object-relational mapping metadata
