+++ lib/Doctrine/ORM/Mapping/UnderscoreNamingStrategy.php
@@ -18,7 +18,7 @@
  * <http://www.doctrine-project.org>.
  */
 
-namespace Doctrine\ORM;
+namespace Doctrine\ORM\Mapping;
 
 /**
  * The default NamingStrategy
@@ -28,18 +28,52 @@ namespace Doctrine\ORM;
  * @since   2.3
  * @author  Fabio B. Silva <fabio.bat.silva@gmail.com>
  */
-class DefaultNamingStrategy implements NamingStrategy
+class UnderscoreNamingStrategy implements NamingStrategy
 {
+    /**
+     * @var string
+     */
+    private $case;
+
+    /**
+     * Underscore naming strategy construct
+     *
+     * @param integer $case CASE_LOWER | CASE_UPPER
+     */
+    public function __construct($case = CASE_LOWER)
+    {
+        $this->case = $case;
+    }
+
+    /**
+     * @return integer
+     */
+    public function getCase()
+    {
+        return $this->case;
+    }
+
+    /**
+     * Sets string case CASE_LOWER | CASE_UPPER
+     * Alphabetic characters converted to lowercase or uppercase
+     * 
+     * @param integer $case
+     */
+    public function setCase($case)
+    {
+        $this->case = $case;
+    }
+
     /**
      * {@inheritdoc}
      */
     public function classToTableName($className)
     {
         if (strpos($className, '\\') !== false) {
-            return substr($className, strrpos($className, '\\') + 1);
+            $className = substr($className, strrpos($className, '\\') + 1);
         }
 
-        return $className;
+        return $this->underscore($className);
     }
 
     /**
@@ -47,7 +81,7 @@ class DefaultNamingStrategy implements NamingStrategy
      */
     public function propertyToColumnName($propertyName)
     {
-        return $propertyName;
+        return $this->underscore($propertyName);
     }
 
     /**
@@ -55,7 +89,7 @@ class DefaultNamingStrategy implements NamingStrategy
      */
     public function referenceColumnName()
     {
-        return 'id';
+        return $this->case === CASE_UPPER ?  'ID' : 'id';
     }
 
     /**
@@ -63,7 +97,7 @@ class DefaultNamingStrategy implements NamingStrategy
      */
     public function joinColumnName($propertyName)
     {
-        return $propertyName . '_' . $this->referenceColumnName();
+        return $this->underscore($propertyName) . '_' . $this->referenceColumnName();
     }
 
     /**
@@ -71,8 +105,7 @@ class DefaultNamingStrategy implements NamingStrategy
      */
     public function joinTableName($sourceEntity, $targetEntity, $propertyName = null)
     {
-        return strtolower($this->classToTableName($sourceEntity) . '_' .
-                $this->classToTableName($targetEntity));
+        return $this->classToTableName($sourceEntity) . '_' . $this->classToTableName($targetEntity);
     }
     
     /**
@@ -80,7 +113,22 @@ class DefaultNamingStrategy implements NamingStrategy
      */
     public function joinKeyColumnName($entityName, $referencedColumnName = null)
     {
-        return strtolower($this->classToTableName($entityName) . '_' .
-                ($referencedColumnName ?: $this->referenceColumnName()));
+        return $this->classToTableName($entityName) . '_' .
+                ($referencedColumnName ?: $this->referenceColumnName());
+    }
+    
+    /**
+     * @param string $string
+     * @return string
+     */
+    private function underscore($string)
+    {
+        $string = preg_replace('/(?<=[a-z])([A-Z])/', '_$1', $string);
+
+        if ($this->case === CASE_UPPER) {
+            return strtoupper($string);
+        }
+
+        return strtolower($string);
     }
 }
